pyproject.toml
[project]
name = "mcp-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gitpython>=3.1.45",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.3",
    "keyboard>=0.13.5",
    "langgraph>=1.0.2",
    "maincontentextractor>=0.0.4",
    "nest-asyncio>=1.6.0",
    "pyaudio>=0.2.14",
    "pydantic>=2.12.4",
    "pygithub>=2.8.1",
    "pyperclip>=1.11.0",
    "python-dotenv>=1.2.1",
    "ratelimit>=2.2.1",
    "requests>=2.32.5",
    "tenacity>=9.1.2",
    "termcolor>=3.2.0",
    "websockets>=15.0.1",
    "groq>=0.29.0",
    "google-genai>=1.45.0",
    "openai>=1.93.0",
    "ollama>=0.5.1",
    "cerebras-cloud-sdk>=1.50.1",
    "anthropic>=0.68.1",
    "mistralai>=1.9.11",
    "pillow>=12.0.0",
]
